requirements.txt
blinker==1.8.2
cachelib==0.13.0
click==8.1.7
Flask==3.0.3
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
msgspec==0.18.6
numpy==2.0.1
opencv-python==4.10.0.84
pillow==10.4.0
scikit-learn==1.5.1
scipy==1.14.0
threadpoolctl==3.5.0
tk==0.1.0
Werkzeug==3.0.3
